gpu: pvr: remove unnecessary udelay from the HW poll loop
At the moment the HW polling loop does a 50 usec delay between
each reading of the HW flag in question. Since this delay is no
worse than just reading the HW flag continuously, get rid of it.

This will reduce the wait time from 50 usec to 25 usec in the
general case.

Signed-off-by: Imre Deak <imre.deak@nokia.com>