gpu: pvr: pdump: rewrite PDumpMemUM()
Now that we have cleared up the streams, we can delay the copy_from_user()
until copying into the streams. This makes PDumpMemUM highly similar to
PDumpMemKM.

So, split out the printing part from PDumpMemKM, and handle parameter
stream dumping separately for PDumpMemKM and PDumpMemUM.

In turn, this allows the removal of the separate buffer which was the
heart of pdump_common.c, which is removed as well.

No functional change.

Signed-off-by: Luc Verhaegen <libv@codethink.co.uk>
Signed-off-by: Imre Deak <imre.deak@nokia.com>