staging: pvr: Makefile: fix to build with v5.3-rc and compile only for enabled ARCH/SOC
after this patch: "kbuild: modpost: read modules.order instead of $(MODVERDIR)/*.mod"
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=ff9b45c55b2659d14420424da0ce4e8aa3dbce28

we need to merge the modules.order files manually so that the new collector
mechanism finds the kernel modules in Phase 2.

In addition we add more ifdefs so that only those kernel modules are built where
CONFIG_ARCH_OMAP or CONFIG_SOC_ is configured.

Signed-off-by: H. Nikolaus Schaller <hns@goldelico.com>