ARM: dts: sun6i: a31s: add sgx gpu child node
We are adding the devicetree binding for the PVR-SGX-544-115 gpu.

This driver is currently under development in the openpvrsgx-devgroup.
Right now the full binding is not figured out, so we provide here a
placeholder. It will be completet as soon as there is a demo available.

The currently used binding that is used during developement is more
complete an was already verifyed by loading the kernelmodule successful.

Signed-off-by: Philipp Rossak <embed3d@gmail.com>