Revert "drm/omap: Fix page fault handling and flush what framebuffe wants flushed"
This reverts commit 33bc438d6d8883d77e37b369fe5144ee9b01fad8.
This causes n900 to fail, and seems to be a wrong fix. With this commit
reverted, we now almost constantly see trails on the left side of the LCD
on droid4 at least with termite on sway.

We now have page fault handling never in use. Let's fix the tests and
use less confusing naming for the helper.

We must flush what framebuffer wants flushed on page flips even if using
page faults like we do with tiler.

Otherwise we have at least droid4 command mode LCD not update a terminal
window as tested with droid4 and sway.

Signed-off-by: Tony Lindgren <tony@atomide.com>